+++ src/Audit.EntityFramework/EntityFrameworkEvent.cs
@@ -1,5 +1,6 @@
 using Newtonsoft.Json;
 using System.Collections.Generic;
+using Audit.Core;
 #if NETSTANDARD1_5 || NETSTANDARD2_0 || NET461
 using Microsoft.EntityFrameworkCore;
 #elif NET45
@@ -7,7 +8,7 @@ using System.Data.Entity;
 #endif
 namespace Audit.EntityFramework
 {
-    public class EntityFrameworkEvent
+    public class EntityFrameworkEvent : IAuditOutput
     {
         [JsonProperty(Order = 1)]
         public string Database { get; set; }
@@ -32,6 +33,9 @@ namespace Audit.EntityFramework
         [JsonIgnore]
         internal DbContext DbContext { get; set; }
 
+        [JsonExtensionData]
+        public Dictionary<string, object> CustomFields { get; set; } = new Dictionary<string, object>();
+
         /// <summary>
         /// Returns the DbContext associated to this event
         /// </summary>
@@ -39,5 +43,22 @@ namespace Audit.EntityFramework
         {
             return DbContext;
         }
+
+        /// <summary>
+        /// Serializes this Entity Framework event as a JSON string
+        /// </summary>
+        public string ToJson()
+        {
+            return JsonConvert.SerializeObject(this, Audit.Core.Configuration.JsonSettings);
+        }
+        /// <summary>
+        /// Parses an Entity Framework event from its JSON string representation.
+        /// </summary>
+        /// <param name="json">JSON string with the Entity Framework event representation.</param>
+        public static EntityFrameworkEvent FromJson(string json)
+        {
+            return JsonConvert.DeserializeObject<EntityFrameworkEvent>(json, Audit.Core.Configuration.JsonSettings);
+        }
+
     }
 }