+++ src/Audit.Mvc/AuditAction.cs
@@ -1,9 +1,10 @@
 using System.Collections.Generic;
+using Audit.Core;
 using Newtonsoft.Json;
 
 namespace Audit.Mvc
 {
-    public class AuditAction
+    public class AuditAction : IAuditOutput
     {
         [JsonProperty(Order = -1, NullValueHandling = NullValueHandling.Ignore)]
         public string TraceId { get; set; }
@@ -47,5 +48,23 @@ namespace Audit.Mvc
         public string RedirectLocation { get; set; }
         [JsonProperty(Order = 999, NullValueHandling = NullValueHandling.Ignore)]
         public string Exception { get; set; }
+        [JsonExtensionData]
+        public Dictionary<string, object> CustomFields { get; set; } = new Dictionary<string, object>();
+
+        /// <summary>
+        /// Serializes this Audit Action as a JSON string
+        /// </summary>
+        public string ToJson()
+        {
+            return JsonConvert.SerializeObject(this, Core.Configuration.JsonSettings);
+        }
+        /// <summary>
+        /// Parses an Audit Action from its JSON string representation.
+        /// </summary>
+        /// <param name="json">JSON string with the Entity Audit Action representation.</param>
+        public static AuditAction FromJson(string json)
+        {
+            return JsonConvert.DeserializeObject<AuditAction>(json, Core.Configuration.JsonSettings);
+        }
     }
 }