+++ src/Audit.NET/AuditEvent.cs
@@ -8,7 +8,7 @@ namespace Audit.Core
     /// <summary>
     /// Represents the output of the audit process
     /// </summary>
-    public class AuditEvent
+    public class AuditEvent : IAuditOutput
     {
         /// <summary>
         /// The enviroment information